+++ src/xyz/oihana/schema/constants/traits/business/DocumentsTrait.php
@@ -3,6 +3,7 @@
 namespace xyz\oihana\schema\constants\traits\business;
 
 use xyz\oihana\schema\constants\traits\business\documents\AdjustmentTrait;
+use xyz\oihana\schema\constants\traits\business\documents\AgingSummaryTrait;
 use xyz\oihana\schema\constants\traits\business\documents\AppliedEcoFeeTrait;
 use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentLineTrait;
 use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentTrait;
@@ -33,6 +34,7 @@ use xyz\oihana\schema\constants\traits\business\documents\TaxDetailTrait;
 trait DocumentsTrait
 {
     use AdjustmentTrait           ,
+        AgingSummaryTrait         ,
         AppliedEcoFeeTrait        ,
         BusinessDocumentLineTrait ,
         BusinessDocumentTrait     ,